+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Delegs.hs
@@ -50,9 +50,9 @@ import Cardano.Ledger.Shelley.LedgerState
     DPState (..),
     RewardAccounts,
     dpsDState,
+    dsUnified,
+    psStakePoolParams,
     rewards,
-    _pParams,
-    _unified,
   )
 import Cardano.Ledger.Shelley.Rules.Delpl (DelplEnv (..), ShelleyDELPL, ShelleyDelplEvent, ShelleyDelplPredFailure)
 import Cardano.Ledger.Shelley.TxBody
@@ -228,14 +228,14 @@ delegsTransition = do
               Map.empty
               wdrls_
           unified' = rewards' UM.⨃ wdrls_'
-      pure $ dpstate {dpsDState = ds {_unified = unified'}}
+      pure $ dpstate {dpsDState = ds {dsUnified = unified'}}
     gamma :|> c -> do
       dpstate' <-
         trans @(ShelleyDELEGS era) $ TRC (env, dpstate, gamma)
 
       let isDelegationRegistered = case c of
             DCertDeleg (Delegate deleg) ->
-              let stPools_ = _pParams $ dpsPState dpstate'
+              let stPools_ = psStakePoolParams $ dpsPState dpstate'
                   targetPool = _delegatee deleg
                in if eval (targetPool ∈ dom stPools_)
                     then Right ()