+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Rules/TestChain.hs
@@ -354,7 +354,7 @@ checkPreservation SourceSignalTarget {source, target, signal} =
             "\n\nUnregistered Treasury MIR total\n",
             show (fold unRegMirTre),
             "\n\nPools Retiring This epoch\n",
-            show (Map.filter (\e -> e == (nesEL . chainNes $ source)) (_retiring . dpsPState . lsDPState $ lsOld)),
+            show (Map.filter (\e -> e == (nesEL . chainNes $ source)) (psRetiring . dpsPState . lsDPState $ lsOld)),
             "\n\ntxs\n"
           ]
             ++ obligationMsgs
@@ -373,7 +373,7 @@ checkPreservation SourceSignalTarget {source, target, signal} =
     ru' = nesRu . chainNes $ source
     lsOld = esLState . nesEs . chainNes $ source
     lsNew = esLState . nesEs . chainNes $ target
-    pools = _pParams . dpsPState . lsDPState $ lsOld
+    pools = psStakePoolParams . dpsPState . lsDPState $ lsOld
     oldRAs = rewards . dpsDState . lsDPState $ lsOld
     newRAs = rewards . dpsDState . lsDPState $ lsNew
 
@@ -390,7 +390,7 @@ checkPreservation SourceSignalTarget {source, target, signal} =
               show obligationDiff
             ]
 
-    mir = _irwd . dpsDState . lsDPState $ lsOld
+    mir = dsIRewards . dpsDState . lsDPState $ lsOld
     isRegistered kh _ = UM.member kh oldRAs
     (regMirRes, unRegMirRes) = Map.partitionWithKey isRegistered (iRReserves mir)
     (regMirTre, unRegMirTre) = Map.partitionWithKey isRegistered (iRTreasury mir)
@@ -563,11 +563,11 @@ potsSumIncreaseByRewardsPerTx SourceSignalTarget {source = chainSt, signal = blo
         { source =
             LedgerState
               UTxOState {utxosUtxo = u, utxosDeposited = d, utxosFees = f}
-              DPState {dpsDState = DState {_unified = umap1}},
+              DPState {dpsDState = DState {dsUnified = umap1}},
           target =
             LedgerState
               UTxOState {utxosUtxo = u', utxosDeposited = d', utxosFees = f'}
-              DPState {dpsDState = DState {_unified = umap2}}
+              DPState {dpsDState = DState {dsUnified = umap2}}
         } =
         (coinBalance u' <+> d' <+> f')
           <-> (coinBalance u <+> d <+> f)
@@ -641,7 +641,7 @@ preserveBalance SourceSignalTarget {source = chainSt, signal = block} =
         LedgerState (UTxOState {utxosUtxo = u'}) _ = ledgerSt'
         txb = tx ^. bodyTxL
         certs = toList (txb ^. certsTxBodyL)
-        pools = _pParams . dpsPState $ dstate
+        pools = psStakePoolParams . dpsPState $ dstate
         created =
           coinBalance u'
             <+> txb ^. feeTxBodyL
@@ -680,7 +680,7 @@ preserveBalanceRestricted SourceSignalTarget {source = chainSt, signal = block} 
         inps === outs
         where
           txb = tx ^. bodyTxL
-          pools = _pParams . dpsPState $ dstate
+          pools = psStakePoolParams . dpsPState $ dstate
           inps =
             coinBalance @era (UTxO (Map.restrictKeys u (txb ^. inputsTxBodyL)))
               <> keyRefunds pp_ txb