+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/UnitTests.hs
@@ -50,8 +50,8 @@ import Cardano.Ledger.Shelley.LedgerState
     LedgerState (..),
     UTxOState (..),
     dpsDState,
+    dsUnified,
     rewards,
-    _unified,
   )
 import Cardano.Ledger.Shelley.PParams
 import Cardano.Ledger.Shelley.Rules
@@ -392,7 +392,7 @@ ledgerState :: LedgerState C
 ledgerState = LedgerState utxoState dpState
 
 addReward :: DPState C_Crypto -> Credential 'Staking C_Crypto -> Coin -> DPState C_Crypto
-addReward dp ra c = dp {dpsDState = ds {_unified = rewards'}}
+addReward dp ra c = dp {dpsDState = ds {dsUnified = rewards'}}
   where
     ds = dpsDState dp
     rewards' = UM.insert ra c (rewards ds)